requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy==1.26.2
fastai==2.7.13
pillow==10.2.0
scikit-image==0.22.0
scikit-learn==1.3.2
opencv-contrib-python==4.9.0.80